Fix warnings in nc_test4

+++ nc_test4/tst_strings.c
@@ -434,7 +434,7 @@ main(int argc, char **argv)
       for (dim_combo = 0; dim_combo < NUM_DIM_COMBOS; dim_combo++)
       {
          char filename[NC_MAX_NAME + 1];
-         int dim_len = dim_combo ? NC_UNLIMITED : DHR_LEN;
+         size_t dim_len = dim_combo ? NC_UNLIMITED : DHR_LEN;
          int expected_unlimdimid = dim_combo ? 0 : -1;
          char *default_fill = ((char *)"");
          char **string_fillp = dim_combo == 3 ? my_string_fill : &default_fill;
@@ -464,7 +464,7 @@ main(int argc, char **argv)
          if (dim_combo)
          {
             size_t start[NDIMS], count[NDIMS] = {1};
-            int counter = 1;
+            size_t counter = 1;
             
             /* Write one record at a time. */
             for (start[0] = 0; start[0] < DHR_LEN; start[0]++)
